# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: ADRO Control API
  version: 0.1.0
  description: Provider-independent Agentic Delivery & Release Orchestrator API.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
- url: "/"
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  "/api/v1/chats":
    get:
      summary: List ordinary project-bound chat sessions
      parameters:
      - name: project_id
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Chat sessions
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
    post:
      summary: Create a durable ordinary chat session
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ChatSessionCreate"
      responses:
        '201':
          description: Chat session created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ChatSession"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
            Idempotency-Replayed:
              "$ref": "#/components/headers/IdempotencyReplayed"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
  "/api/v1/chats/{id}":
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      summary: Read chat messages and harness context status
      responses:
        '200':
          description: Chat session
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: adro_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional for POST, PUT, and PATCH. Scope is tenant + method + path.
        Reuse with the same query, content type, and body replays the first completed
        response; reuse with different input returns 409 idempotency_key_conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 255
  schemas:
...
    ChatSessionCreate:
      type: object
      required:
      - workspace_id
      - title
      properties:
        workspace_id:
          type: string
        project_id:
          type: string
        title:
          type: string
        budget_tokens:
          type: integer
          minimum: 0
        auto_compaction:
          type: boolean
        compaction_threshold:
          type: number
          minimum: 0
          maximum: 1
        compaction_retain_tail:
          type: integer
          minimum: 0
    ChatSession:
      type: object
      required:
      - id
      - workspace_id
      - title
      - harness_session_id
      - status
      properties:
        id:
          type: string
        workspace_id:
          type: string
        project_id:
          type: string
        title:
          type: string
        harness_session_id:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    JsonObject:
      type: object
      description: Versioned JSON resource or page. Resource-specific fields remain
        backward-compatible within /api/v1.
      additionalProperties: true
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        error_code:
          type: string
        request_id:
          type: string
      additionalProperties: true
...
  headers:
    RequestID:
      description: Stable correlation identifier for this HTTP request.
      schema:
        type: string
    IdempotencyReplayed:
      description: True when the status, headers, and body were replayed from the
        first completed request.
      schema:
        type: string
        const: 'true'